+++ src/components/pages/hero.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react'
 import { classnames } from '@/utils'
-import Content from '@/components/UI/content'
+import Content from '@/components/ui/content'
 import styles from './Hero.module.css'
-import Button from '@/components/UI/button'
-import Socials from '@/components/UI/socials/socials'
+import Button from '@/components/ui/button'
+import Socials from '@/components/ui/socials/socials'
 
 const Hero: FC = () => {
   return (
